+++ PROD_Wakgood/Config/DefaultEngine.ini
@@ -45,10 +45,11 @@ AppliedDefaultGraphicsPerformance=Maximum
 r.GenerateMeshDistanceFields=True
 r.DynamicGlobalIlluminationMethod=1
 r.ReflectionMethod=1
-r.Shadow.Virtual.Enable=1
+r.Shadow.Virtual.Enable=0
 r.DefaultFeature.AutoExposure.ExtendDefaultLuminanceRange=True
 r.DefaultFeature.LocalExposure.HighlightContrastScale=0.8
 r.DefaultFeature.LocalExposure.ShadowContrastScale=0.8
+r.DefaultFeature.MotionBlur=False
 
 [/Script/WorldPartitionEditor.WorldPartitionEditorSettings]
 CommandletClass=Class'/Script/UnrealEd.WorldPartitionConvertCommandlet'
