+++ PROD_Wakgood/Source/PROD_Wakgood/Character/WakCharacterBase.cpp
@@ -2,27 +2,86 @@
 
 // Wak Header
 #include "Character/WakHealthComponent.h"
+#include "Interaction/WakInteractionBase.h"
 
 // Unreal Header
 #include "AbilitySystemComponent.h"
+#include "EnhancedInputComponent.h"
+#include "EnhancedInputSubsystems.h"
 #include "Components/CapsuleComponent.h"
 #include "GameFramework/CharacterMovementComponent.h"
+#include "Components/BoxComponent.h"
+#include "GameFramework/SpringArmComponent.h"
+#include "Camera/CameraComponent.h"
 
 AWakCharacterBase::AWakCharacterBase(const FObjectInitializer& ObjectInitializer)
 {
 	PrimaryActorTick.bCanEverTick = false;
 	PrimaryActorTick.bStartWithTickEnabled = false;
 
+	// Configure character movement
+	GetCharacterMovement()->bOrientRotationToMovement = true;	
+	GetCharacterMovement()->RotationRate = FRotator(0.0f, 1000.0f, 0.0f);
+	GetCharacterMovement()->JumpZVelocity = 700.f;
+	GetCharacterMovement()->AirControl = 0.35f;
+	GetCharacterMovement()->MaxWalkSpeed = 500.f;
+	
+	CameraBoom = CreateDefaultSubobject<USpringArmComponent>(TEXT("CameraBoom"));
+	CameraBoom->SetupAttachment(RootComponent);
+	CameraBoom->SetWorldRotation(FRotator(-30.f, 0.f, 0.f));
+	CameraBoom->SetUsingAbsoluteRotation(true);
+	CameraBoom->TargetArmLength = 600.0f;
+	
+	FollowCamera = CreateDefaultSubobject<UCameraComponent>(TEXT("FollowCamera"));
+	FollowCamera->SetupAttachment(CameraBoom, USpringArmComponent::SocketName);
+	FollowCamera->bUsePawnControlRotation = false;
+
 	HealthComponent = CreateDefaultSubobject<UWakHealthComponent>(TEXT("HealthComponent"));
 	HealthComponent->OnDeathStarted.AddDynamic(this, &ThisClass::OnDeathStarted);
 	HealthComponent->OnDeathFinished.AddDynamic(this, &ThisClass::OnDeathFinished);
 
+	InteractionBox = CreateDefaultSubobject<UBoxComponent>(TEXT("InteractionBox"));
+	InteractionBox->SetupAttachment(RootComponent);
+	InteractionBox->SetCollisionProfileName(TEXT("OverlapAll"));
+}
+
+void AWakCharacterBase::BeginPlay()
+{
+	Super::BeginPlay();
+
+	if (APlayerController* PlayerController = Cast<APlayerController>(Controller))
+	{
+		if (UEnhancedInputLocalPlayerSubsystem* Subsystem = ULocalPlayer::GetSubsystem<UEnhancedInputLocalPlayerSubsystem>(PlayerController->GetLocalPlayer()))
+		{
+			Subsystem->AddMappingContext(DefaultMappingContext, 0);
+		}
+	}
+	
+	InteractionBox->OnComponentBeginOverlap.AddDynamic(this, &AWakCharacterBase::OnBeginOverlap);
+	InteractionBox->OnComponentEndOverlap.AddDynamic(this, &AWakCharacterBase::OnEndOverlap);
 }
 
 void AWakCharacterBase::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
 {
 	Super::SetupPlayerInputComponent(PlayerInputComponent);
 
+	// Set up action bindings
+	if (UEnhancedInputComponent* EnhancedInputComponent = Cast<UEnhancedInputComponent>(PlayerInputComponent)) {
+		
+		// Jumping
+		EnhancedInputComponent->BindAction(JumpAction, ETriggerEvent::Started, this, &ACharacter::Jump);
+		EnhancedInputComponent->BindAction(JumpAction, ETriggerEvent::Completed, this, &ACharacter::StopJumping);
+
+		// Moving
+		EnhancedInputComponent->BindAction(MoveAction, ETriggerEvent::Triggered, this, &AWakCharacterBase::Move);
+
+		// interacting
+		EnhancedInputComponent->BindAction(InteractionAction, ETriggerEvent::Started, this, &AWakCharacterBase::OnInteract);
+	}
+	else
+	{
+		UE_LOG(LogTemp, Error, TEXT("'%s' Failed to find an Enhanced Input component! This template is built to use the Enhanced Input system. If you intend to use the legacy system, then you will need to update this C++ file."), *GetNameSafe(this));
+	}
 }
 
 UAbilitySystemComponent* AWakCharacterBase::GetAbilitySystemComponent() const
@@ -81,6 +140,81 @@ void AWakCharacterBase::UninitAndDestroy()
 	}
 
 	SetActorHiddenInGame(true);
+}
+
+void AWakCharacterBase::Move(const FInputActionValue& Value)
+{
+	FVector2D MovementVector = Value.Get<FVector2D>();
+
+	if (Controller != nullptr)
+	{
+		const FRotator Rotation = Controller->GetControlRotation();
+		const FRotator YawRotation(Rotation.Pitch, 0, 0);
+		const FVector RightDirection = FRotationMatrix(YawRotation).GetUnitAxis(EAxis::Y);
+		AddMovementInput(RightDirection, MovementVector.X);
+	}
+}
+
+void AWakCharacterBase::OnBeginOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor,
+                                       UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult)
+{
+	GEngine->AddOnScreenDebugMessage(0, 5, FColor::Red, FString::Printf(TEXT("Character : OnBeginOverlap Called")));
+
+	if (OtherActor && (OtherActor != this) && OtherComp)
+	{
+		if(!InteractionTarget && GetWorld())
+		{
+			InteractionTarget = Cast<AInteractionBase>(OtherActor);
+		}
+	}
+}
+
+void AWakCharacterBase::OnEndOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor,
+	UPrimitiveComponent* OtherComp, int32 OtherBodyIndex)
+{
+	GEngine->AddOnScreenDebugMessage(0, 5, FColor::Red, FString::Printf(TEXT("Character : OnEndOverlap Called")));
+	if (InteractionTarget)
+	{
+		InteractionTarget = nullptr;
+	}
+}
+
+void AWakCharacterBase::OnInteract()
+{
+	if (Controller != nullptr)
+	{
+		GEngine->AddOnScreenDebugMessage(0, 5, FColor::Red, FString::Printf(TEXT("Interaction Input Detected")));
+		SwitchInteractInput();
+		if(InteractionTarget != nullptr)
+		{
+			Interface = Cast<IInteractionInterface>(InteractionTarget);
+			if (Interface)
+			{
+				Interface->Interaction(this);
+			}
+		}
+		else
+		{
+			GEngine->AddOnScreenDebugMessage(0, 5, FColor::Red, FString::Printf(TEXT("null interactionTarget")));
+		}
+	}
+}
+
+void AWakCharacterBase::SwitchInteractInput()
+{
+	if (bIsInteractInput)
+	{
+		bIsInteractInput = false;
+	}
+	else
+	{
+		bIsInteractInput = true;
+	}
+}
+
+bool AWakCharacterBase::GetIsInteractInput() const
+{
+	return bIsInteractInput;
 }
