+++ PROD_Wakgood/Source/PROD_Wakgood/Game/WakLevelScript.cpp
@@ -4,7 +4,6 @@
 #include "WakLevelScript.h"
 
 #include "WakMainGameInstance.h"
-#include "PROD_Wakgood/TP_ThirdPerson/WakThirdPersonCharacter.h"
 
 // Sets default values
 ALevelScript::ALevelScript()
