--- PROD_Wakgood/Source/PROD_Wakgood/TP_ThirdPerson/TP_ThirdPerson.h
@@ -1,5 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
