--- PROD_Wakgood/Source/PROD_Wakgood/TP_ThirdPerson/TP_ThirdPersonGameMode.cpp
@@ -1,14 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#include "TP_ThirdPersonGameMode.h"
-#include "UObject/ConstructorHelpers.h"
-
-ATP_ThirdPersonGameMode::ATP_ThirdPersonGameMode()
-{
-	// set default pawn class to our Blueprinted character
-	static ConstructorHelpers::FClassFinder<APawn> PlayerPawnBPClass(TEXT("/Game/ThirdPerson/Blueprints/BP_ThirdPersonCharacter.BP_ThirdPersonCharacter"));
-	if (PlayerPawnBPClass.Class != NULL)
-	{
-		DefaultPawnClass = PlayerPawnBPClass.Class;
-	}
-}
