--- PROD_Wakgood/Source/PROD_Wakgood/TP_ThirdPerson/TP_ThirdPersonGameMode.h
@@ -1,17 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "WakThirdPersonCharacter.h"
-#include "GameFramework/GameModeBase.h"
-#include "TP_ThirdPersonGameMode.generated.h"
-
-UCLASS(minimalapi)
-class ATP_ThirdPersonGameMode : public AGameModeBase
-{
-	GENERATED_BODY()
-
-public:
-	ATP_ThirdPersonGameMode();
-};
