--- PROD_Wakgood/Source/PROD_Wakgood/TP_ThirdPerson/WakThirdPersonCharacter.h
@@ -1,90 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "InputActionValue.h"
-#include "PROD_Wakgood/Interaction/WakInteractionInterface.h"
-#include "PROD_Wakgood/Interaction/WakInteractionBase.h"
-#include "GameFramework/Character.h"
-#include "Logging/LogMacros.h"
-#include "WakThirdPersonCharacter.generated.h"
-
-class USpringArmComponent;
-class UCameraComponent;
-class UInputMappingContext;
-class UInputAction;
-class UBoxComponent;
-class AInteractionBase;
-struct FInputActionValue;
-
-DECLARE_LOG_CATEGORY_EXTERN(LogTemplateCharacter, Log, All);
-
-UCLASS(config=Game)
-class AWakThirdPersonCharacter : public ACharacter, public IInteractionInterface
-{
-	GENERATED_BODY()
-
-	/** Camera boom positioning the camera behind the character */
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera, meta = (AllowPrivateAccess = "true"))
-	USpringArmComponent* CameraBoom;
-
-	/** Follow camera */
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera, meta = (AllowPrivateAccess = "true"))
-	UCameraComponent* FollowCamera;
-	
-	/** MappingContext */
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
-	UInputMappingContext* DefaultMappingContext;
-
-	/** Jump Input Action */
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
-	UInputAction* JumpAction;
-
-	/** Move Input Action */
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
-	UInputAction* MoveAction;
-
-	// interaction action
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = Input, meta = (AllowPrivateAccess = "true"))
-	UInputAction* InteractionAction;
-
-	UPROPERTY(EditAnywhere)
-	UBoxComponent* InteractionBox;
-
-private:
-	IInteractionInterface* Interface = nullptr;
-
-public:
-	AWakThirdPersonCharacter();
-	
-
-protected:
-	void Move(const FInputActionValue& Value);
-	
-	// interaction
-	UFUNCTION()
-	void OnBeginOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor,
-		UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
-	UFUNCTION()
-	void OnEndOverlap(UPrimitiveComponent* OverlappedComp, AActor* OtherActor,
-		UPrimitiveComponent* OtherComp, int32 OtherBodyIndex);
-	
-	AInteractionBase* InteractionTarget;
-	
-	void OnInteract();
-	bool bIsInteractInput = false;
-
-public:
-	void switchInteractInput();
-	bool getIsInteractInput();
-	
-protected:
-	virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;	
-	virtual void BeginPlay();
-
-public:
-	FORCEINLINE class USpringArmComponent* GetCameraBoom() const { return CameraBoom; }
-	FORCEINLINE class UCameraComponent* GetFollowCamera() const { return FollowCamera; }
-
-};
